Transfer retention api test and fts to Kibana Management
